refactor: move the signs editor out of AdminPanel
It was already a self-contained component, SignsView, just defined inside a
2573-line file alongside everything else a GM can do. Moving it was a lift
rather than a carve: the five constants it used were used by nothing else in the
file, and its props are unchanged, so App still owns the state and the API calls
still go to the same routes.

Renamed to SignEditor now that it stands on its own — SignsView read as "the
view AdminPanel switches to", which is no longer what it is.

AdminPanel drops from 2573 lines to 2151.

The eight rotation tests render AdminPanel and drive this editor, so they now
exercise it through the module and pass unchanged, which is the evidence the
move preserved behaviour rather than only compiling.

backend/ is still untouched and the 24 signs API tests still pass, so signs
already placed are unaffected.
