+++ frontend/src/modules/signs/index.ts
@@ -8,4 +8,5 @@
 
 export { Signs } from './components/Signs';
 export { AutoSignage } from './components/AutoSignage';
+export { SignEditor } from './components/SignEditor';
 export type { SignData, SignLine } from './components/Signs';
